+++ Controller/QueueController.php
@@ -3,9 +3,12 @@
 namespace Dtc\QueueBundle\Controller;
 
 use Dtc\QueueBundle\Model\Worker;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\StreamedResponse;
 
 class QueueController extends Controller
 {
@@ -48,6 +51,33 @@ class QueueController extends Controller
     }
 
     /**
+     * @Route("/archive", name="dtc_queue_archive")
+     * @Method({"POST"})
+     */
+    public function archiveAction(Request $request)
+    {
+        $workerName = $request->get('workerName');
+        $methodName = $request->get('method');
+
+        $jobManager = $this->get('dtc_queue.job_manager');
+        $callback = function ($count) {
+            echo json_encode(['count' => $count]);
+            echo "\n";
+            flush();
+        };
+
+        return new StreamedResponse(function () use ($jobManager, $callback, $workerName, $methodName) {
+            $total = $jobManager->countLiveJobs($workerName, $methodName);
+            echo json_encode(['total' => $total]);
+            echo "\n";
+            flush();
+            if ($total > 0) {
+                $jobManager->archiveAllJobs($workerName, $methodName, $callback);
+            }
+        });
+    }
+
+    /**
      * List jobs in system by default.
      *
      * @Template("@DtcQueue/Queue/jobs.html.twig")
@@ -59,7 +89,29 @@ class QueueController extends Controller
         $managerType = $this->container->getParameter('dtc_queue.default_manager');
         $rendererFactory = $this->get('dtc_grid.renderer.factory');
         $renderer = $rendererFactory->create('datatables');
-        $gridSource = $this->get('dtc_queue.grid_source.live_jobs.'.('mongodb' === $managerType ? 'odm' : $managerType));
+        $gridSource = $this->get('dtc_queue.grid_source.jobs_waiting.'.('mongodb' === $managerType ? 'odm' : $managerType));
+        $renderer->bind($gridSource);
+        $params = $renderer->getParams();
+        $this->addCssJs($params);
+
+        $params['worker_methods'] = $this->get('dtc_queue.job_manager')->getWorkersAndMethods();
+
+        return $params;
+    }
+
+    /**
+     * List jobs in system by default.
+     *
+     * @Template("@DtcQueue/Queue/jobs_running.html.twig")
+     * @Route("/jobs_running", name="dtc_queue_jobs_running")
+     */
+    public function runningJobsAction()
+    {
+        $this->validateManagerType('dtc_queue.default_manager');
+        $managerType = $this->container->getParameter('dtc_queue.default_manager');
+        $rendererFactory = $this->get('dtc_grid.renderer.factory');
+        $renderer = $rendererFactory->create('datatables');
+        $gridSource = $this->get('dtc_queue.grid_source.jobs_running.'.('mongodb' === $managerType ? 'odm' : $managerType));
         $renderer->bind($gridSource);
         $params = $renderer->getParams();
         $this->addCssJs($params);