+++ DependencyInjection/Compiler/WorkerCompilerPass.php
@@ -34,8 +34,13 @@ class WorkerCompilerPass implements CompilerPassInterface
 
         $managerType = $this->getRunManagerType($container);
         $jobTimingManagerType = $this->getJobTimingManagerType($container);
-        $container->setParameter('dtc_queue.class_job_timing', $this->getClass($container, $jobTimingManagerType, 'job_timing',
-            'JobTiming', JobTiming::class));
+        $container->setParameter('dtc_queue.class_job_timing', $this->getClass(
+            $container,
+            $jobTimingManagerType,
+            'job_timing',
+            'JobTiming',
+            JobTiming::class
+        ));
         $container->setParameter('dtc_queue.class_run', $this->getClass($container, $managerType, 'run', 'Run', Run::class));
         $container->setParameter('dtc_queue.class_run_archive', $this->getClass($container, $managerType, 'run_archive', 'RunArchive', Run::class));
 
@@ -46,6 +51,7 @@ class WorkerCompilerPass implements CompilerPassInterface
             $eventDispatcher->addMethodCall('addSubscriber', [$eventSubscriber]);
         }
         $this->setupDoctrineManagers($container);
+        $this->addLiveJobs($container);
     }
 
     protected function setupAlias(ContainerBuilder $container, $defaultManagerType, $type)
@@ -103,6 +109,9 @@ class WorkerCompilerPass implements CompilerPassInterface
         }
     }
 
+    /**
+     * @param ContainerBuilder $container
+     */
     protected function setupDoctrineManagers(ContainerBuilder $container)
     {
         $documentManager = $container->getParameter('dtc_queue.document_manager');
@@ -110,7 +119,6 @@ class WorkerCompilerPass implements CompilerPassInterface
         $odmManager = "doctrine_mongodb.odm.{$documentManager}_document_manager";
         if ($container->has($odmManager)) {
             $container->setAlias('dtc_queue.document_manager', $odmManager);
-            GridSourceCompilerPass::addGridSource($container, 'dtc_queue.grid_source.live_jobs.odm');
         }
 
         $entityManager = $container->getParameter('dtc_queue.entity_manager');
@@ -118,7 +126,22 @@ class WorkerCompilerPass implements CompilerPassInterface
         $ormManager = "doctrine.orm.{$entityManager}_entity_manager";
         if ($container->has($ormManager)) {
             $container->setAlias('dtc_queue.entity_manager', $ormManager);
-            GridSourceCompilerPass::addGridSource($container, 'dtc_queue.grid_source.live_jobs.orm');
+        }
+    }
+
+    /**
+     * @param ContainerBuilder $container
+     */
+    protected function addLiveJobs(ContainerBuilder $container)
+    {
+        $jobReflection = new \ReflectionClass($container->getParameter('dtc_queue.class_job'));
+        if ($jobReflection->isInstance(new \Dtc\QueueBundle\Document\Job())) {
+            GridSourceCompilerPass::addGridSource($container, 'dtc_queue.grid_source.jobs_waiting.odm');
+            GridSourceCompilerPass::addGridSource($container, 'dtc_queue.grid_source.jobs_running.odm');
+        }
+        if ($jobReflection->isInstance(new \Dtc\QueueBundle\Entity\Job())) {
+            GridSourceCompilerPass::addGridSource($container, 'dtc_queue.grid_source.jobs_waiting.orm');
+            GridSourceCompilerPass::addGridSource($container, 'dtc_queue.grid_source.jobs_running.orm');
         }
     }