+++ Tests/Doctrine/BaseJobManagerTest.php
@@ -229,6 +229,106 @@ abstract class BaseJobManagerTest extends BaseBaseJobManagerTest
         self::assertEquals($id, $result->getId());
     }
 
+    public function testCountLiveJobs()
+    {
+        /** @var JobManager|\Dtc\QueueBundle\ORM\JobManager $jobManager */
+        $jobManager = self::$jobManager;
+
+        while ($job = $jobManager->getJob()) {
+            $jobManager->deleteJob($job);
+        }
+
+        $this->getJob();
+
+        $count = $jobManager->countLiveJobs();
+        self::assertEquals(1, $count);
+
+        $this->getJob();
+
+        $count = $jobManager->countLiveJobs();
+        self::assertEquals(2, $count);
+
+        $this->getJob();
+
+        $count = $jobManager->countLiveJobs();
+        self::assertEquals(3, $count);
+
+        $count = $jobManager->countLiveJobs('asdf');
+        self::assertEquals(0, $count);
+
+        $count = $jobManager->countLiveJobs('fibonacci');
+        self::assertEquals(3, $count);
+
+        $count = $jobManager->countLiveJobs('fibonacci', 'test');
+        self::assertEquals(0, $count);
+
+        $count = $jobManager->countLiveJobs('fibonacci', 'fibonacci');
+        self::assertEquals(3, $count);
+
+        while ($job = $jobManager->getJob()) {
+            $jobManager->deleteJob($job);
+        }
+    }
+
+    public function testArchiveAllJobs()
+    {
+        /** @var JobManager|\Dtc\QueueBundle\ORM\JobManager $jobManager */
+        $jobManager = self::$jobManager;
+
+        while ($job = $jobManager->getJob()) {
+            $jobManager->deleteJob($job);
+        }
+
+        $this->getJob();
+
+        $count = $jobManager->countLiveJobs();
+        $archiveCount = $this->runCountQuery($jobManager->getJobArchiveClass());
+        self::assertEquals(1, $count);
+        $allCount = $this->runCountQuery($jobManager->getJobClass());
+        $counter = 0;
+        $countJobs = function ($count) use (&$counter) {
+            $counter += $count;
+        };
+        $jobManager->archiveAllJobs(null, null, $countJobs);
+        self::assertEquals(0, $jobManager->countLiveJobs());
+        self::assertEquals($allCount - 1, $this->runCountQuery($jobManager->getJobClass()));
+        self::assertEquals($archiveCount + 1, $this->runCountQuery($jobManager->getJobArchiveClass()));
+        self::assertEquals(1, $counter);
+
+        $this->getJob();
+        $this->getJob();
+
+        $count = $jobManager->countLiveJobs();
+        self::assertEquals(2, $count);
+        $archiveCount = $this->runCountQuery($jobManager->getJobArchiveClass());
+        $counter = 0;
+        $jobManager->archiveAllJobs('fibonacci', null, $countJobs);
+        self::assertEquals(0, $jobManager->countLiveJobs());
+        self::assertEquals(2, $counter);
+        self::assertEquals($archiveCount + 2, $this->runCountQuery($jobManager->getJobArchiveClass()));
+
+        $this->getJob();
+        $this->getJob();
+
+        $count = $jobManager->countLiveJobs();
+        self::assertEquals(2, $count);
+
+        $jobManager->archiveAllJobs('fibonacc', null, $countJobs);
+        self::assertEquals(2, $jobManager->countLiveJobs());
+
+        $jobManager->archiveAllJobs('fibonacci', 'fibo', $countJobs);
+        self::assertEquals(2, $jobManager->countLiveJobs());
+
+        $jobManager->archiveAllJobs('fibonacci', 'fibonacci', $countJobs);
+        self::assertEquals(0, $jobManager->countLiveJobs());
+
+        while ($job = $jobManager->getJob()) {
+            $jobManager->deleteJob($job);
+        }
+    }
+
+    abstract protected function runCountQuery($class);
+
     public function testResetErroneousJobs()
     {
         /** @var JobManager|\Dtc\QueueBundle\ORM\JobManager $jobManager */